requirements.txt
numpy
pandas 
tqdm
matplotlib
scikit-learn
tensorflow-cpu
tensorflow-hub
opencv-python
seaborn
